requirements.txt
Django==1.10.4

dj-database-url==0.4.2
django-cleanup==0.4.2
django-compressor==2.1.1
django-debug-toolbar==1.7
django-extensions==1.7.8
django-imagekit==3.3
django-libsass==0.7
djangorestframework==3.5.3

libsass==0.12.3
mistune==0.7.4
pilkit==1.1.13
Pillow==3.4.2

psycopg2==2.6.2
python-decouple==3.0
uWSGI==2.0.15

factory-boy==2.8.1
